# csharp: Optimize the performance of the Hash algorithm and reduce allocations.