<!-- Please make sure the target branch is right. In most case, the target branch should be `master`. -->

### Which Issue(s) This PR Fixes

<!-- Please ensure that the related issue has already been created, and [link this pull request to that issue using keywords](<https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword>) to ensure automatic closure. -->

Just optimize performance.


### Brief Description

<!-- Write a brief description for your pull request to help the maintainer understand the reasons behind your changes. -->

Conducting benchmark tests on the previous and current methods using Benchmark, the speed has increased nearly threefold, and memory usage has also been reduced by 50%.



| Method       | Mean     | Error   | StdDev  | Gen0   | Allocated |
|------------- |---------:|--------:|--------:|-------:|----------:|
| AfterGetMd5  | 162.7 ns | 0.47 ns | 0.44 ns | 0.0095 |     168 B |
| AfterGetSha1 | 151.9 ns | 0.34 ns | 0.30 ns | 0.0114 |     200 B |
| GetMd5       | 468.8 ns | 1.25 ns | 1.17 ns | 0.0238 |     416 B |
| GetSha1      | 510.4 ns | 1.35 ns | 1.26 ns | 0.0267 |     472 B |


### How Did You Test This Change?

Existing unit tests cover these changes.

<!-- In order to ensure the code quality of Apache RocketMQ, we expect every pull request to have undergone thorough testing. -->